Remove mention of changing permissions
Though we do currently allow changes to the permissions for an app user property, the permissions are not ever made visible to the admin and we have deliberately avoided mentioning them in the API until now.  I've removed the mention from my change, so all I'm adding is the part about nullability.  (I will also open a ticket suggesting we disallow changing the permissions via the API until we are prepared to admit their existence.)